Add files via upload
This project was difficult. I looked at the Slack comments, tried to reverse engineer the solutions, and applied some of the solutions to my TomeRater project. The method / class concepts made overall sense in the lessons, but I was not clear on how to apply the method concepts to this project. After review, could you please send the solution code? I would like to study it. Thank you.